requirements.txt
anyio==3.6.1
beanie==1.11.9
click==8.1.3
dnspython==2.2.1
fastapi==0.81.0
h11==0.13.0
httptools==0.4.0
idna==3.3
motor==3.0.0
multidict==6.0.2
numpy==1.23.2
pandas==1.4.4
pydantic==1.10.1
pymongo==4.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
toml==0.10.2
typing_extensions==4.3.0
uvicorn==0.18.3
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
yarl==1.8.1
